reccworker: include pid in default bot_id
Currently the default bot_id is composed only of the hostname. This
means that if a user wishes to host more than one bot per host, they are
required to define a more sophisticated identification scheme when
invoking reccworker at the command line. With this change, multiple
instaces per host functions correctly and uniquely out of the box.

Update the logging so that when the bot comes up in verbose mode it logs
its bot_id which will help debugging.