Remove whitespace stripping from conf file parser
Currently when parsing a recc conf file we will strip all whitespace
wherever it occurs in the string. This leads to issues with, for
example, customisation of the message displayed when a user is
unauthenticated: any spaces in the string are removed.

It appears that this functionality is only a convenience and not relied
upon anywhere. We are making the parser more strict and correct.