Log Operation ID
Log the operation ID (when available) given by the execution service.

Currently troubleshooting the life of a job submitted by recc is tricky
if the execution service does not log the Action ID. Even then, it's
better to work with Operation IDs. recc currently logs the action ID.

In verbose mode, log only once the operation ID that is given to recc by
the execution service.