Improve ACTION_CACHE and CAS_SERVER fallback messages
Currently when you run recc in verbose mode without setting either the
CAS or Action Cache (AC) endpoints it will write a message saying that
the Execution Service (ES) endpoint is being used for the CAS and then
say that the Execution Service endpoint is being used for the
ActionCache:

```
$ RECC_VERBOSE=1 ./recc/opt/bb/bin/recc echo hi
Warning: no RECC_SERVER environment variable specified.Using default server (localhost:8085)
No RECC_CAS_SERVER environment variable specified.Using the same as RECC_SERVER (localhost:8085)
No RECC_ACTION_CACHE_SERVER environment variable specified.Using the same as RECC_SERVER (localhost:8085)
```

In practice the code was actually doing the right thing and using the
CAS endpoint for the AC but the error message was slightly wrong and the
code slightly convoluted.

I have simplified the logic so that the fallback order is first to match
between the CAS and the AC and then fallback to ES if neither has a URI
specifieid. That way recc will be seen to be falling back to the CAS for
the AC since that is the logical place for the AC to live per
discussion[1].

Now the result shows that the AC is falling back to `RECC_CAS_SERVER`:

```
$ RECC_VERBOSE=1 ./recc g++ foo
Warning: no RECC_SERVER environment variable specified. Using default server (localhost:8085)
No RECC_CAS_SERVER environment variable specified. Using the same as RECC_SERVER (localhost:8085)
No RECC_ACTION_CACHE_SERVER environment variable specified. Using the same as RECC_CAS_SERVER (localhost:8085)
```

Additionally, if the AC is specified use that for the CAS over the ES
since that makes the most sense (CAS and AC live together usually).

Added tests to check this functionality.

[1] https://github.com/bazelbuild/remote-apis/pull/76/files