Move upload missing blobs metric to cover all uploads
There is a case where if the blob exceeds the batch size we upload it without
capturing the metric. Move the blob upload metric to capture all logic in
batchUpdateBlobs function.

Uploading a large number of blobs, or large blobs, may result in more than one
rpc since these will be cut up into smaller batches. In the future we may want
to count the number of batchUpdateBlobs calls as well as the time for each
individual one.

Signed-off-by: Ed Baunton <ebaunton1@bloomberg.net>