Add RECC_CACHE_ONLY option
Due to how environment variables are resolved, and channels are
configured, it's difficult to add a cache-only mode without adding
another option.

This is because there is no way to distinguish between RECC_SERVER
unset and set to the default after variables are resolved, and when
forced in, everything assumes a remote channel is created, even if not
used, so we need an address for RECC_SERVER (or a _lot_ of reworking
of the internals).

Instead, let's create a RECC_CACHE_ONLY option. This gives users a bit
more control over this feature, too, so it's a nice to have.